+++ src/main/java/pl/plajer/villagedefense/arena/ArenaRegistry.java
@@ -23,6 +23,7 @@ import java.util.List;
 import java.util.logging.Level;
 
 import org.bukkit.Bukkit;
+import org.bukkit.Difficulty;
 import org.bukkit.configuration.ConfigurationSection;
 import org.bukkit.configuration.file.FileConfiguration;
 import org.bukkit.entity.Player;
@@ -151,7 +152,13 @@ public class ArenaRegistry {
       arena.setStartLocation(LocationSerializer.getLocation(config.getString(key + "Startlocation", "world,364.0,63.0,-72.0,0.0,0.0")));
       arena.setEndLocation(LocationSerializer.getLocation(config.getString(key + "Endlocation", "world,364.0,63.0,-72.0,0.0,0.0")));
       ArenaUtils.setWorld(arena);
-
+      if (arena.getStartLocation().getWorld().getDifficulty() == Difficulty.PEACEFUL){
+        Bukkit.getConsoleSender().sendMessage(plugin.getChatManager().colorMessage(Messages.VALIDATOR_INVALID_ARENA_CONFIGURATION).replace("%arena%", id).replace("%error%", "THERE IS A WRONG " +
+            "DIFFICULTY -> SET IT TO ANOTHER ONE THAN PEACEFUL"));
+        arena.setReady(false);
+        ArenaRegistry.registerArena(arena);
+        continue;
+      }
       if (!config.getBoolean(key + "isdone", false)) {
         Bukkit.getConsoleSender().sendMessage(plugin.getChatManager().colorMessage(Messages.VALIDATOR_INVALID_ARENA_CONFIGURATION).replace("%arena%", id).replace("%error%", "NOT VALIDATED"));
         arena.setReady(false);