+++ CHANGELOG.md
@@ -18,6 +18,7 @@ That's all, matcher will stop when detects next line started with `###` match
 * Fixed SpectatorItemEvent not loaded, now spectator items will work on click
 * Fixed special items stage loading is always lobby
 * Fixed leave special items not working
+* Added difficulty validator
 
 ### 4.2.2 Release (14.06.2019 - 13.07.2019)
 * Fixed setup menu completed/not completed values were buggy